Changed module structure of blog metric chart builders
