--- app/services/builders/blog_metric_chart/technology_blog_post_count.rb
@@ -1,9 +1,0 @@
-module Builders
-  module BlogMetricChart
-    class TechnologyBlogPostCount < Builders::BlogMetricChart::Base
-      def metric_name
-        Metric.names[:blog_post_count]
-      end
-    end
-  end
-end
