--- app/services/builders/blog_metric_chart/technology_visits.rb
@@ -1,9 +1,0 @@
-module Builders
-  module BlogMetricChart
-    class TechnologyVisits < Builders::BlogMetricChart::Base
-      def metric_name
-        Metric.names[:blog_visits]
-      end
-    end
-  end
-end
