--- app/services/builders/blog_metric_chart/technology_visits_growth_mom.rb
@@ -1,9 +1,0 @@
-module Builders
-  module BlogMetricChart
-    class TechnologyVisitsGrowthMom < Builders::BlogMetricChart::GrowthMonthOverMonth
-      def metric_name
-        Metric.names[:blog_visits]
-      end
-    end
-  end
-end
