+++ app/services/builders/metric_chart/blog/technology_visits.rb
@@ -1,0 +1,11 @@
+module Builders
+  module MetricChart
+    module Blog
+      class TechnologyVisits < Builders::MetricChart::Blog::Base
+        def metric_name
+          Metric.names[:blog_visits]
+        end
+      end
+    end
+  end
+end
