+++ app/services/builders/metric_chart/blog/technology_visits_growth_mom.rb
@@ -1,0 +1,11 @@
+module Builders
+  module MetricChart
+    module Blog
+      class TechnologyVisitsGrowthMom < Builders::MetricChart::Blog::GrowthMonthOverMonth
+        def metric_name
+          Metric.names[:blog_visits]
+        end
+      end
+    end
+  end
+end
